requirements.txt
flask
openai
python-dotenv
python-telegram-bot==20.3
langchain
langchain-community
tiktoken
pytz
